bitbot/requirements.txt
flask
flask_cors
flask-limiter
flask_sqlalchemy
pandas
redis
fpdf2
schedule
datetime
PyPDF2
locust
pika
numpy
cryptography
celery